requirements.txt
flask
flask-cors
PyPDF2
python-docx
